requirements.txt
beautifulsoup4==4.8.2
blaze==0.10.1
bs4==0.0.1
click==7.0
cycler==0.10.0
cython==0.29.14
datashape==0.5.2
decorator==4.4.1
flask-cors==3.0.8
flask==1.1.1
h5py==2.10.0
itsdangerous==1.1.0
jinja2==2.10.3
joblib==0.14.1
kiwisolver==1.1.0
lxml==4.4.2
markupsafe==1.1.1
matplotlib==3.1.2
multipledispatch==0.6.0
networkx==2.4
nltk==3.4.5
numpy==1.18.1
odo==0.5.0
pandas==0.23.4
psutil==5.6.7
py4j==0.10.7
pyarrow==0.15.1
pyjnius==1.1.1
pymagnitude==0.1.120
pyparsing==2.4.6
pyspark==2.4.4
python-dateutil==2.8.1
pytz==2019.3
sarge==0.1.5.post0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
soupsieve==1.9.5
sqlalchemy==1.3.12
toolz==0.10.0
tqdm==4.41.1
trectools==0.0.37
werkzeug==0.16.0

